fix: Change user_id to BigInteger for video progress
Date.now() generates timestamps that exceed 32-bit integer limit.
Changed user_id column to BigInteger to support large user IDs.

+++ app/models/video_progress.py
@@ -1,5 +1,5 @@
 # app/models/video_progress.py
-from sqlalchemy import Column, Integer, String, Float, DateTime, UniqueConstraint
+from sqlalchemy import Column, Integer, BigInteger, String, Float, DateTime, UniqueConstraint
 from sqlalchemy.sql import func
 from app.db.base import Base
 
@@ -12,7 +12,7 @@ class UserVideoProgress(Base):
     __tablename__ = "user_progress"
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    user_id = Column(Integer, nullable=False, index=True)
+    user_id = Column(BigInteger, nullable=False, index=True)
     video_id = Column(String(50), nullable=False, index=True)
     seconds_accumulated = Column(Float, default=0.0, nullable=False)
     last_updated = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now())
